+++ src/app/components/cluster-host-table.js
@@ -1,11 +1,22 @@
+/**
+ * Table with cluster hosts
+ *
+ * @module components/cluster-host-table
+ * @author Jakub Liput
+ * @copyright (C) 2017-2018 ACK CYFRONET AGH
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
 import { readOnly } from '@ember/object/computed';
 import { observer, computed } from '@ember/object';
 import {
   InvokeActionMixin,
 } from 'ember-invoke-action';
 import BasicTable from 'onedata-gui-common/components/basic-table';
-
+import I18n from 'onedata-gui-common/mixins/components/i18n';
 import { validator, buildValidations } from 'ember-cp-validations';
+import notImplementedReject from 'onedata-gui-common/utils/not-implemented-reject';
+import { scheduleOnce } from '@ember/runloop';
 
 const roles = ['database', 'clusterWorker', 'clusterManager'];
 
@@ -54,11 +65,19 @@ let Validations = buildValidations(generateColumnValidations(roles));
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 export default BasicTable.extend(
+  I18n,
   InvokeActionMixin,
   hostColumnComputedProperties(roles),
   Validations, {
     tagName: 'table',
     classNames: ['cluster-host-table', 'table', 'table-striped', 'dropdown-table-rows'],
+
+    i18nPrefix: 'components.clusterHostTable',
+
+    /**
+     * @virtual optional
+     */
+    removeHost: notImplementedReject,
 
     /**
      * To inject.
@@ -78,13 +97,27 @@ export default BasicTable.extend(
     // TODO make/use valid properties for each column
     // databaseHostsValid: computed.readOnly('validations.attrs.databaseHosts.isValid'),
 
+    /**
+     * @type {Ember.ComputedProperty<boolean>}
+     */
+    removeHostAvailable: computed('removeHost', function () {
+      return this.get('removeHost') !== notImplementedReject;
+    }),
+
     tableValidChanged: observer('allValid', function () {
       this.invokeAction('allValidChanged', this.get('allValid') === true);
+    }),
+
+    hostsChanged: observer('hosts.[]', function () {
+      scheduleOnce('afterRender', this, '_reinitializeBasictable');
     }),
 
     init() {
       this._super(...arguments);
       this.tableValidChanged();
+
+      // enable observers
+      this.get('hosts.[]');
     },
 
     actions: {
@@ -94,6 +127,10 @@ export default BasicTable.extend(
 
       primaryClusterManagerChanged(hostname, isSet) {
         this.invokeAction('primaryClusterManagerChanged', isSet ? hostname : null);
+      },
+
+      removeHost(hostname) {
+        return this.get('removeHost')(hostname);
       },
     },
   }