+++ src/app/components/create-admin-form.js
@@ -26,9 +26,29 @@ export default Component.extend(I18n, {
 
   i18nPrefix: 'components.createAdminForm.',
 
+  /**
+   * @virtual 
+   * @type {Function}
+   */
   registerStarted: notImplementedWarm,
+
+  /**
+   * @virtual 
+   * @type {Function}
+   */
   registerFailure: notImplementedWarm,
+
+  /**
+   * @virtual 
+   * @type {Function}
+   */
   registerSuccess: notImplementedWarm,
+
+  /**
+   * @virtual 
+   * @type {Function}
+   */
+  back: notImplementedWarm,
 
   /**
    * @type {string}
@@ -52,9 +72,19 @@ export default Component.extend(I18n, {
   passwordEntered: false,
   confirmEntered: false,
   confirmTyped: false,
+  usernameEntered: false,
 
   passwordsMatch: computed('password', 'confirmPassword', function () {
     return this.get('password') === this.get('confirmPassword');
+  }),
+
+  usernameInvalidMessage: computed('username', function () {
+    const username = this.get('username');
+    if (username.length < 2) {
+      return 'tooShort';
+    } else {
+      return null;
+    }
   }),
 
   confirmInvalid: computed('confirmEntered', 'passwordsMatch', function () {
@@ -81,26 +111,35 @@ export default Component.extend(I18n, {
     'username',
     'password',
     'passwordInvalidMessage',
+    'usernameInvalidMessage',
     function () {
       const {
         confirmValid,
         username,
         password,
         passwordInvalidMessage,
+        usernameInvalidMessage,
       } = this.getProperties(
         'confirmValid',
         'username',
         'password',
         'passwordInvalidMessage',
+        'usernameInvalidMessage',
       );
 
       return username &&
         password &&
         !passwordInvalidMessage &&
+        !usernameInvalidMessage &&
         confirmValid;
     }),
 
   didInsertElement() {
+    this.$('.add-user-username').on('focusout', () => {
+      if (!this.get('usernameEntered')) {
+        safeExec(this, 'set', 'usernameEntered', true);
+      }
+    });
     this.$('.password').on('focusout', () => {
       if (!this.get('passwordEntered')) {
         safeExec(this, 'set', 'passwordEntered', true);
@@ -157,6 +196,10 @@ export default Component.extend(I18n, {
       } else {
         return Promise.reject();
       }
+    },
+
+    back() {
+      this.get('back')(...arguments);
     },
   },
 });