+++ src/app/locales/en/components/new-cluster-installation.js
@@ -7,4 +7,9 @@ export default {
     'your Zone, you need to properly set up DNS zone delegation. ' +
     'For more information, see',
   domainLink: 'the documentation',
+  addNewHost: 'Add new host',
+  newHostnameExample: 'Enter hostname or IP address',
+  save: 'Add host',
+  cancel: 'Cancel',
+  addingNewHost: 'adding new host',
 };