+++ src/app/services/cluster-manager.js
@@ -373,11 +373,12 @@ export default Service.extend({
   },
 
   /**
+   * @param {string} type
    * @return {Promise} resolves with Array.{ hostname: string }
    */
-  getHosts() {
+  getHosts(type = 'known') {
     return new Promise((resolve, reject) => {
-      let gettingHostNames = this.getHostNames();
+      let gettingHostNames = this.getHostNames(type);
 
       gettingHostNames.then(({ data: hostnames }) => {
         // TODO more info
@@ -395,14 +396,26 @@ export default Service.extend({
     });
   },
 
+  /**
+   * @param {string} type one of: known, cluster
+   * @returns {Promise}
+   */
   getHostNames() {
     let onepanelServer = this.get('onepanelServer');
-    return new Promise((resolve, reject) => {
-      let gettingClusterHosts = onepanelServer.requestValidData(
-        'onepanel',
-        'getClusterHosts',
-      );
-      gettingClusterHosts.then(resolve, reject);
-    });
+    return onepanelServer.requestValidData(
+      'onepanel',
+      'getClusterHosts',
+    );
+  },
+
+  /**
+   * @param {string} address hostname or IP address
+   * @returns {Promise<Onepanel.KnownHost>}
+   */
+  addKnownHost(address) {
+    return this.get('onepanelServer').request(
+      'onepanel',
+      'addClusterHost', { address }
+    ).then(({ data }) => data);
   },
 });